refactor: 💡 extract SearchInput component
